refactor: extract exceptions into separate file
