+++ parser/args.js
@@ -1,45 +1,8 @@
 // @flow
 /* eslint-disable no-console */
+import { InvalidArgument, InvalidConfig } from '../error';
+
 export type Args = string | Array<string>
-
-class BaseError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = this.constructor.name;
-
-    if (typeof Error.captureStackTrace === 'function') {
-      Error.captureStackTrace(this, this.constructor);
-    } else {
-      this.stack = new Error(message).stack;
-    }
-  }
-}
-
-export class InvalidArgument extends BaseError {
-  config: Args;
-  args: Array<string>;
-
-  constructor(
-    message: string,
-    config: Args,
-    args: Array<string>,
-  ) {
-    super(message);
-    this.message = message;
-    this.config = config;
-    this.args = args;
-  }
-}
-
-export class InvalidConfig extends BaseError {
-  config: mixed;
-
-  constructor(message: string, config: mixed) {
-    super(message);
-    this.message = message;
-    this.config = config;
-  }
-}
 
 function validateArguments(config, args) {
   if (typeof config === 'object' && config !== null && config.length) {
